requirements.txt
python-dotenv
requests
groq
PyMuPDF
json
tiktoken